requirements.txt
altgraph==0.17
AnyQt==0.0.11
autopep8==1.5.6
certifi==2020.12.5
confapp==1.1.11
cycler==0.10.0
future==0.18.2
importlib-metadata==3.7.3
keyboard==0.13.5
kiwisolver==1.3.1
matplotlib==3.3.4
natsort==7.1.1
numpy==1.19.5
opencv-python==3.4.5.20
pefile==2019.4.18
Pillow==8.1.2
playsound==1.2.2
pycodestyle==2.7.0
PyForms-GUI==4.904.152
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
PySimpleGUI==4.38.0
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
QScintilla==2.12.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
visvis==1.13.0
wincertstore==0.2
zipp==3.4.1
